# swagger.yaml
openapi: 3.0.3
info:
  title: Web API
  description: |-
  license:
    name: API WEB 1.0
    url: "http://localhost:8989/api/v0"
  version: 1.0.11
servers:
  - url: http://localhost:8989/api/v0
    description: Local server
...
paths:
...
  /user:
    get:
      tags:
        - User
      description: get user
      operationId: getUser
      summary: Get user by id
      parameters:
        - name: id
          in: query
          description: id user
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
            application/xml:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid ID supplied
        "404":
          description: User not found
        "422":
          description: Validation exception
      # security:
      #   - petstore_auth:
      #       - write:pets
      #       - read:pets

    put:
      tags:
        - User
      description: update user
      operationId: update user
      summary: Update user
      requestBody:
        description: Update user
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                username:
                  type: string
                  example: "admin"
      
                fullname:
                  type: string
                  example: "admin"
                email:
                  type: string
                  example: "nguen@gmail.com"
               
                phone:
                  type: integer
                  example: 12345
                sex:
                  type: string
                  example: "NAM"
                avatar:
                  type: string
                  example: "admin"
                birthday:
                  type: string
                  format: date
                  example: "2024-04-14"
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "422":
          description: Validation exception
      # security:
      #   - petstore_auth:
      #       - write:pets
      #       - read:pets
    delete:
      tags:
        - User
      description: delete user
      operationId: deleteUser
      summary: Delete user
      parameters:
        - name: id
          in: query
          description: id user
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: User not found
        "422":
          description: Validation exception
  /category:
    post:
      tags:
        - Category
      summary: Add category
      description: add category
      operationId: addCategory
      requestBody:
        description: Add category
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "admin"
                user_id:
                  type: integer
                  example: "11"
                image:
                  type: string
                  example: "https://example.com/images/web-development.jpg"
              required:
                - name
          application/xml:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "admin"
              required:
                - name
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "admin"
              required:
                - name
        required: true
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "422":
          description: Validation exception
      # security:
      #   - petstore_auth:
      #       - write:pets
      #       - read:pets
    get:
      tags:
        - Category
      description: get all categories
      summary: Get all categories
      operationId: getAllCategories
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                            name:
                              type: string
                            user_id:
                              type: integer
                            popular:
                              type: integer
                            slug:
                              type: string
                            image:
                              type: string
                            updatedDate:
                              type: string
                              format: date-time
            
        "400":
          description: Invalid ID supplied
        "422":
          description: Validation exception
      # security:
      #   - petstore_auth:
      #       - write:pets
      #       - read:pets
    put:
      tags:
        - Category
      description: update category
      summary: Update category
      operationId: updateCategory
      requestBody:
        description: Update category
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                name:
                  type: string
                  example: "web landingpage"
                user_id:
                  type: integer
                  example: "11"
                image:
                  type: string
                  example: "https://example.com/images/web-development.jpg"
              required:
                - id
                - name
          application/xml:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                name:
                  type: string
                  example: "admin"
              required:
                - id
                - name
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                name:
                  type: string
                  example: "admin"
              required:
                - id
                - name
        required: true
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "422":
          description: Validation exception
      # security:
      #   - petstore_auth:
      #       - write:pets
      #       - read:pets
    delete:
      tags:
        - Category
      description: delete category
      summary: Delete category
      parameters:
        - name: id
          in: query
          description: id category
          required: true
          schema:
            type: integer

      operationId: deleteCategory
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "422":
          description: Validation exception
      # security:
      #   - petstore_auth:
      #       - write:pets
      #       - read:pets
  /product:
    post:
      tags:
        - Product
      summary: Add product
      description: add product
      operationId: addProduct
      requestBody:
        description: Add product
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
          application/xml:
            schema:
              $ref: "#/components/schemas/ProductInput"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "422":
          description: Validation exception
      # security:
      #   - petstore_auth:
      #       - write:pets
      #       - read:pets
    put:
      tags:
        - Product
      summary: Update product
      description: update product
      operationId: updateProduct
      requestBody:
        description: Update product
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
          application/xml:
            schema:
              $ref: "#/components/schemas/ProductInput"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/ProductInput"
        required: true
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "422":
          description: Validation exception
      # security:
      #   - petstore_auth:
      #       - write:pets
      #       - read:pets

    delete:
      tags:
        - Product
      summary: Delete product
      description: delete product
      operationId: deleteProduct
      parameters:
        - name: id
          in: query
          description: id product
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "422":
          description: Validation exception
      # security:
      #   - petstore_auth:
      #       - write:pets
      #       - read:pets
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        username:
          type: string
          example: theUser
        status_id:
          type: integer
          format: int64
          example: 10
        role_id:
          type: integer
          format: int64
          example: 10
        balance:
          type: integer
          format: int64
          example: 10
        fullname:
          type: string
          example: James
        email:
          type: string
          example: john@email.com
        password:
          type: string
          example: "12345"
        phone:
          type: integer
          example: 12345
        sex:
          type: string
          example: "NAM"
        avatar:
          type: string
          example: James
        points:
          type: integer
          format: int64
          example: 10
        qr_admin:
          type: object
          example: {}
        birthday:
          type: string
          format: date
          example: "2024-04-14"
        isOnline:
          type: boolean
          example: true
        offlineAt:
          type: string
          format: date
          example: "2024-04-14"
        referrer_id:
          type: integer
          format: int64
          example: 10
        referral_code:
          type: string
          example: "12345"
        created_at:
          type: string
          format: date
          example: "2024-04-14"
        updated_at:
          type: string
          format: date
          example: "2024-04-14"
      xml:
        name: user
...
    ProductInput:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        user_id:
          type: integer
          format: int64
          example: 5
        popular:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: "dien thoai"
        price:
          type: integer
          format: int64
          example: 99000
        url_Demo:
          type: string
          example: "https://example.com/demo"
        category:
          type: array
          items:
            type: string
          example: ["WebDevelopment", "JavaScript"]
        description:
          type: string
          example: "This is a product description"
        url_Download:
          type: string
          example: "https://example.com/download"
        technology:
          type: array
          items:
            type: string
          example: ["nodejs", "JavaScript"]
      required:
        - user_id
        - name
        - price
        - description
        - url_Download
...